Move `use_heuristic` into `bids.py`
